pyproject.toml
[project]
name = "techscreen"
version = "0.0.0"
description = "TechScreen — internal AI technical interview system (backend)"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115,<0.120",
  "uvicorn[standard]>=0.32,<0.36",
  "structlog>=24.4,<25",
  "pyyaml>=6.0,<7",
  "google-genai>=0.5,<1",
  "google-api-core>=2.20,<3",
  "tenacity>=9.0,<10",
  "pydantic-settings>=2.6,<3",
]

[dependency-groups]
dev = [
  "ruff>=0.7,<1",
  "mypy>=1.13,<2",
  "pytest>=8.3,<9",
  "httpx>=0.27,<0.29",
  "types-PyYAML>=6.0,<7",
  "pytest-asyncio>=0.23,<1",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [".venv", "node_modules", "app/frontend"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/backend/tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
namespace_packages = true
mypy_path = "."
# Target directory is passed on the CLI (see contracts/dev-commands.md Command 1);
# declaring `files` here would make an empty target error because mypy treats an
# explicit `files` list as "must find these". Leaving it implicit lets the CLI
# argument drive discovery so an empty `app/backend/` still exits 0.

# google-genai 0.x ships without a `py.typed` marker; treat it as untyped
# at the module boundary. The wrapper module (`app/backend/llm/_real_backend.py`)
# is the only consumer and is itself fully typed at the public surface.
[[tool.mypy.overrides]]
module = ["google.genai", "google.genai.*"]
ignore_missing_imports = true
